requirements.txt
MetaTrader5==5.0.36
telethon==1.25.4
python-dotenv==0.21.0
aiogram==2.14.3